requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.18
pydantic==2.11.7
pyyaml==6.0.2
httpx==0.28.1
pytest==8.4.1
